requirements.txt
streamlit
pillow
google-generativeai
python-dotenv
